Change to (home-made) wkhtmltopdf-static
The Debian wkhtmltopdf package will not run without an X11 server.
There is a version that will, but it relies on being linked with a
patched QT library, and so is not available through the usual
Debian package mechanism. [1]

Therefore we have made a custom Debian package containing a statically
linked binary of wkhtmltopdf, linked (statically) against this
patched QT. [2]

1. http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=580226
2. http://debian.mysociety.org/dists/squeeze/main/binary-amd64/wkhtmltopdf-static-0.11.0-rc1.deb