Load all raw emails for testing
Previously there was just one raw email, and the test code relied
on that fact. Generalise it to handle multiple raw emails.

This change causes a number of tests to fail, because it exposes
failures that should have happened when the second raw email was
added but were masked by the fact that the text of this second
raw email was never loaded. These failures will be fixed in the
next commit.