Load all fixtures for all tests
The ad hoc specification of fixtures has been an ongoing source
of bugs in the tests. The straw that broke the camel’s back is
that 7c6eb09 requires the fixtures to be loaded in order (i.e.
children before their parents), and it would have been a painful
process to reorder all the dozens of different fixture lists,
but the test system ought to be more reliable this way.