Specify attachment ordering
So it turns out that issue #408 is a real and potentially nasty
bug. The code was assuming that attachments are returned in the
order they were created, which is a) not guaranteed to be the case,
and b) sometimes actually not the case, hence the occasional test
failures.

This trivial patch corrects the issue by sorting attachments by
id. The effect of this is to return them in the order they were
added to the database, since all were added using the same sequence
in a single database session.

Fixes #408.