Test should not depend on nondeterministic order
Issue #409 seems to be another problem of the same type as #408,
though with the important difference that the bug in this case
is in the test suite rather than the code under test.

RequestMailer sends alert emails warning of overdue requests.
However it does not specify the order that these messages are
sent in, but runs over the overdue requests in whatever order
they are returned by a database query (that does not have an
order by clause). Therefore it is not safe for the test code
to assume that the alert mails will have been sent in a particular
order: just as with #408 it seems that they were *usually* sent
in the order assumed by the test code, but occasionally not --
which would result in sporadic test failures.

Closes #409.