Bundler fix for rails-post-deploy
Ordinarily we would expect simply to run "bundle install" here.
However, at the time of writing there is a bug in bundler that
causes gems from github to be rebuilt every time bundle install
is run, which makes the process very extremely slow in our case
because Xapian takes a long time to build. Running
"bundle exec bundle install" is a workaround for this bug.

However clearly one cannot run bundle exec till the bundle has
initially been installed, so we use a flag file .bundler-has-run
to indicate whether we are doing an initial install.