Atom feed of request events
We need not only new requests, but new outgoing correspondence
of any sort.

The idea is that this feed will contain any event that would have
triggered an email to be sent to the public body, so can be used
as an alternative, equivalent way to stay up-to-date with happenings
on WDTK (or the Alaveteli installation of choice).