Fix another test
Is there actually any point in tests of this sort? It is not testing
that anything works as intended, only that a particular SQL string is
passed to the database -- and surely the only way that string could
have been discovered for test purposes was essentially to copy it
from the code under test.

It seems to me that the only real function of a test of this sort
is to make it more difficult to modify the code under test.