script/plugin no longer exists in Rails 3
Instead we should use "bundle exec rails plugin".  The
"-r rails-3" is to make sure that we get the Rails 3 version
of the theme, i.e. one that uses Rails.root instead of
RAILS_ROOT.  (Unfortunately, it doesn't work specifying an
annotated tag with "-r", so we use a branch name instead,
say, of "use-alaveteli-0.10".)