Fix attachment downloading for attachments that have file extensions
Fixes #912

The routing in Rails 3 by default extracts anything that looks
like a file extension at the end of the URL and adds it as a
format parameter.  This meant that for the get_attachment and
get_attachment_as_html views, the filename parameter would not
include the file extension, and you'll end up with the error:

  please use same filename as original file has,
  display: 'foo.doc' old_display: 'foo.doc' original: 'foo'

For particular routes, we can suppress this with
:format => false so this commit adds that to both get_attachment
and get_attachment_as_html.