Add rake tasks for checking the parsing of a random sample of emails
For the upgrade from Rails 2 to Rails 3, we want to check that
existing emails are parsed correctly.  One of these tasks
(temp:random_attachments_hexdigests) is for dumping a CVS file
of details about each attachment, and its hexdigest, and a list
of the randomly selected raw emails.  (It's intended that you
run this on an old Rails 2 install of Alaveteli.  The other task
(temp:recompute_attachments_hexdigests) is intended to be run
on a Rails 3 install of Alaveteli, with the previous files as
input, to check that the same results are obtained.