Add tests for TNEF attachments that should be handled
These two cases were ignored previously, and we need to make sure
that they still are under the switch from TMail to Mail.

One TNEF attachment is a heavily truncated one from a real example
from Alaveteli that has no personal data in it.  The other is
an example from the tests in the distribution of the tnef
package for Ubuntu 1.4.9-1 - it's an HTML version of the US
constitution.