Remove an unnecessary use of 'which' in backticks
The use of backticks in `which tnef` means that which(1) is
invoked under /bin/sh, and returns the first executable
matching 'tnef' in the default PATH for /bin/sh.  In this
usage, however, Popen also uses /bin/sh to execute the
command, which would run that same executable if just a
bare 'tnef' were used.  In summary, I can't see any reason
for the convolution of:

   IO.popen("#{`which tnef`.chomp} -K -C #{dir}", "wb") do |f|

... over just:

   IO.popen("tnef -K -C #{dir}", "wb") do |f|

... so switch to the latter.