Add a test to check that anything in the MIME epilogue is ignored
There is currently a difference in behaviour in the parsing
of nested MIME multipart attachments between the Mail and
TMail based backends.  This commit adds a test that will
pass if the behaviour is the same as the the old (TMail-based)
version, which I believe is correct according to RFC 1521.
The example email has a PNG attachment after the final MIME
boundary, and the RFC says that anything after the final
boundary ("the epilogue") should be ignored.