Add a test for a missing final MIME boundary
If there is a missing final MIME boundary, the behaviour
of Alaveteli with the TMail backend was to still parse
the attachment, but with the new code it currently throws
an exception.  This commit adds a test that asserts that
the attachment should be parsed despite the email being
malformed in this way.