Cope with emails with a missing final MIME boundary
The Mail gem deals with multipart messages that look as if
they should have 1 part but are missing the final MIME boundary,
by make the parts list empty and setting part.body to the
text of the email.  Rather than throwing an exception in this
case, we just pretend that part is text/plain and return it, so
that the page doesn't error and we still have a chance of some
useful text being displayed.

Note that we haven't investigated yet the case of emails that
have more than one start boundary, but no final boundary.

Fixes #921