Make an attachment filename in test fixtures more difficult
This commit changes 'hello.txt' to 'hello-world.txt' in the
incoming-request-two-same-name.email fixture.  The reasoning
for this change is that if there are no special characters in
the filename then filename and display_filename will be the
same and the tests won't pick up any confusion between the
two.

The test requests to :get_attachment and
:get_attachment_as_html should get the display_filename rather
than filename.