xapian:rebuild_index: fix for "SSL error: decryption failed or bad record mac"
The connection to the database doesn't survive a fork, but can
be re-established if we disconnect beforehand, and use
establish_connection in the child after the fork.

rebuild_index with safe_rebuild=true is only called as a rake task
(i.e. in its own process) so this disconnection and reconnection
shouldn't affect the running of the site.