Check that display_filename matches URL part number or fallback
If the display_filename of the attachment found from the
URL part number doesn't match the passed in display_filename
then the email may have been reparsed, causing a reordering.
In that case, look to see if there is another attachment that
uniquely matches that filename, and, if so, return that other
attachment.  If no matching uniquely matching filename is
found, redirect to the incoming message, rather than
returning a 404.