Improve calculation of PublicBody statistics columns
On PublicBody, we don't need to update info_requests_count
because that's already done with :counter_cache.  On the
other hand, info_requests_successful_count and
info_requests_not_held_count can't be updated easily with
counter_cache (since they need conditions to be attached).
Instead we update them in post_save and post_destroy,
as suggested here:

   http://blog.douglasfshearer.com/post/17495285851/custom-counter-cache-with-conditions

This also adds tests to ensure that the
after_(save|destroy) callbacks are called and that they
modify the counts correctly.