Serve admin.js as a static file
admin.js was being generated from a view so that the config variable
ADMIN_PUBLIC_URL could be used to direct AJAX requests in the admin
interface correctly.  ADMIN_PUBLIC_URL was necessary for
WhatDoTheyKnow when the admin interface was proxied over SSL, while
the rest of the site was non-SSL.  Now that the whole site is served
over SSL, however, the proxy for the admin interface is not longer
used and ADMIN_PUBLIC_URL / ADMIN_BASE_URL have been removed.

As a result, admin.js is now being generated from a Rails view
unnecessarily, and at some performance cost.  This commit moves
admin.coffee and admin.js to the public directory, so they will
be served more efficiently.

In addition, this commit removes admin.coffee / admin.js from
public/javascripts/ where there was a second (and different) version
that now appears to be unused.