Tidy up SQL query strings
In the list method of PublicBodyController, capitalize
the SQL keywords consistenly and don't include extraneous
whitespace in the strings.  (On the former point,
previously only some were - the convention is a matter of
some debate, but in this case the you editor's not doing
to do syntax highlighting of SQL so having the keywords
capitalized helps its readibility, I think.)

I wouldn't normally do this kind of cosmetic tidying up,
since it affects the tidiness of diffs and merges, but in
this case I'm going to change all these lines in the
next commit anyway, so that reasoning doesn't apply.