Don't display duplicate public bodies with the fallback
This introduces some raw SQL statement for the fallback case, but we
can't see an easy way to avoid that in this case.

This commit also adds some tests that assert the sorting and
non-duplication properties of the listing.

Thanks to Louise Crow for working out the SQL expression for
falling back to the default locale.