Add install scripts for Alaveteli
site-specific-install.sh will be called by our generic site
install script in commonlib/bin/install-site.sh

These scripts assume that you have a new installation of
Debian squeeze or Ubuntu precise and then will set up:

 - Alaveteli running in development mode with the Thin web
   server behing nginx

 - The cron jobs that are required for the site to work.

 - A basic Postfix configuration for sending and receiving
   mail.

We also will use this script for generating new AMIs
(Amazon Machin Images) for Alaveteli.

The general.yml configuration file will be created if it
doesn't exist, but if there is an existing copy it won't be
overwritten, so it should be safe to customize that file and
then re-run the install script.