Add package lists for Ubuntu precise and Debian squeeze
The config/packages file cannot be used as a list of packages
to give to apt-get install - its format is designed to be used
as build-dependencies of a fake Debian package.  These
distribution- and release-specific new files, on the other
hand, provide a list of packages that should satisfy the
config/packages requirements on Ubuntu precise or Debian
squeeze.  The generic site install file will use files like
these on the matching distribution and release.