Update the calculation of the numerator for percentage statistics
We have changed the denominator of the proportion-based statistics
to only include requests that are both visible and not
'awaiting_description'.  This would mean, however, that the numerator
could be larger than the denominator.  This commit updates the
calculation of those statistics to also exclude any hidden or
unclassified requests.