Update globalize3 repo.
This is the commit where @henares' pull request was merged. This is a quick fix, but I think we want to see if we can use the globalize 3.0.1 gem with minimal fixes. It causes tests to fail at the moment, I think because we still have translated attributes in our public_bodies table (and not null constraints at the databse level on at least the name attribute).