Use upstream globalize.
We're trying to switch to an upstream version of globalize, which now
has Henare's change to allowed blank values without them being
converted to nil - however, there is another intermediate change that
was introduced which means that the name is only written back to the
original table if the locale matches the default locale.  This is
described in more depth here:

   globalize/globalize#183

This change broke the make_single_language_example helper, but this
commit fixes that by also setting the default locale before creating
the example public bodies.