Move themes from vendor/plugins to lib/themes
These are essentially required in exactly the same way as before, but
from lib/themes rather than vendor/plugins.  This is the simplest
possible change in order make the themes work outside vendor/plugins,
I think, but it's not necessarily ideal.  It would be worth considering
whether these should be changed to Rails engines, as described here:

  http://guides.rubyonrails.org/engines.html