Add timestamps to public_body_translation fixtures.
Between rails 3.2.0 and 4.0.1rc4, timestamps were constrained to be
non-null. As globalize creates the public_body_translation table with
timestamps, this means there are errors in the tests when trying to
insert the fixture data, which doesn't have timestamps. These errors
were only appearing on Travis because locally the test database is
cloned from the development structure rather than being created by
running the migrations.