Try to uninstall the old theme from vendor/plugins and lib/themes
The theme install task would fail if there's an old theme present
in vendor/plugins, since it doesn't try to uninstall the plugin from
that location, only the new location.  Then when the install.rb
in the new plugin runs, it'll complain that there's a
public/alavetelitheme symlink already present.

This commit changes themes:install to try to uninstall the plugin
from both locations.