Move disclosure log to common part of form.
It isn't in the list of attributes that is translated, so was causing a
mass assignment error when the code in public_body.translated_versions=
was attempting to assign it. There's a possible underlying inconsistency
here between the treatment of publication_scheme and disclosure_log -
one is translated and one isn't. I've ticketed that separately (#1264)