Use a better subject line than "Fake Response" for uploaded responses
The subject line wasn't being explicitly specified in
RequestMailer.fake_response and a default of "Fake Response" (somehow
based on the method name) was being used instead; this subject line
would sometimes be visible to users and authorities and caused alarm
to them.  Instead, we use the standard subject lines for followup
emails for the InfoRequest.

Fixes #1105.