Create development environment with Vagrant
Share the host git checkout in to the VM so that a developer can use
their preferred environment while running the application in a
production-like environment.

Use the --dev option of install-site.sh so that the existing git checkout is
used.

The gems are installed outside of the app directory in development
because VirtualBox's directory shares can be slow when the directory
contains an excessive amount of files. Installing gems in to this
directory results in ~15K additional files, which is beyond what
VirtualBox can reasonably be expected to handle