Use default log dir and work around for mySociety
Really logs should be in /var/log/APP_NAME/log, with APP_ROOT/log being
symlinked to that location as a convenience. We're not at the point of
being able to do this yet, but what we shouldn't be doing is moving the
log directory somewhere completely non-standard.

This commit creates the log directory in the expected place, and uses a
symlink to adhere to the mySociety non-standard setup