Add a single primary key column on PublicBodyCategoryLink
Without this, cascaded deletes of category -> heading relationships
can't work, because ActiveRecord can't figure out how to select the
right record from the join table, as it can't use the composite key
we had in place for that.

This commit adds a new auto-incremented id field to be the primary
key, and relegates the previous composite key to just being a unique
index.

Closes #1963