Add a “Back to search results” link
Store the search parameters in the flash when a search is made via the
select_authority_path page. Stores the parameters for both POST and
AJAX typeahead searches.

The presence of stored search params renders the link on the
PublicBodyController#show template.

“keep”s the search params in PublicBodyController#show so that if the
user clicks the browser’s back button the “Back to search results” link
can still be rendered on the next search result they click.

“keep”s all flash keys in ServicesController#other_country_message
as it’s called through AJAX and ends up sweeping the flash. [1]

[1] More details about this:
    http://mikenaberezny.com/2007/09/08/keep-the-flash-and-test-it-too/