Add pry to development gems
Pry is awesome. The main reason to add it to the Gemfile is to use
`binding.pry` in views and controllers. TL;DR is that it halts execution
and opens a REPL for you to dig around at the current variables at the
point of the binding. See Railscast #280 [1] for a quick overview.

0.9 seems to be the last version to support Ruby 1.8

[1] http://railscasts.com/episodes/280-pry-with-rails