Handle a child that cannot be terminated.
Basic handling that is the same as for other process errors - just write to
stderr. Add TODO for allowing the calling code to specify an error stream
to allow other handling. In different circumstances, the calling code might want
to raise an exception, write to the log, or do something else. At the
moment this code assumes a cron context and writes to stderr, which will
cause an email to be sent from the cron process, but that's not always the
context the calling code is being run in. So we may be missing errors that are, for
example, being written to stderr during the request cycle.

In fact, this code may be called from the same
place with different contexts (i.e. as part of a cron reindexing task and
in the request cycle), so at the moment some substantial refactoring of the
calling code would be required to make sure calls to this code handle errors
appropriately for each context.