select text after the box is positioned
Some devices (IOS) jump the screen to
the position of a form input when it is selected.

This causes the screen to scroll to the bottom
of the page when the 'link to this' icon is clicked,
because the browser scrolls before the box is positioned
on screen.

This commit changes the order of steps in
script for the 'link to this' widget, so that the
input is selected *after* the box is positioned.

This stops the broswer from jumping
to the bottom of the page.

Tested in Chrome, Safari and IOS simulator on Mac OSX.

Fixes #2264
https://github.com/mysociety/alaveteli/issues/2264