Check for blank values, not just an empty string
On my machine the integration test "Editing a Public Body can add a
translation for multiple locales" (http://git.io/vvv9t) was failing
because when it submitted the form to update one locale it created
empty records for all locales resulting in the test
(http://git.io/vvv93) for a nil locale to fail.

It was doing this because when it submitted the form it included a
notes field with a single newline character and so this method
evaluated as false and therefore a largely blank translation was
created.