Change link to browse other requests to one that has requests
When you're starting a request to an authority that has no existing
requests we display a handy link to "Browse other requests for
examples of how to word your request". However this currently sends
you to a page that has no requests (a blank request search page).

This change sends you to the list of successful requests as these are
most likely to have wording that will help the requester.