Use existing attribute
`last_event_time` is set as a result of calling
`InfoRequest.find_in_state`. This commit adds an explicit accessor that
converts it to a `DateTime` and uses it in the view, preventing an extra
query for each `InfoRequest` in the result set.

Before:
Completed 200 OK in 14908.9ms (Views: 14625.3ms | ActiveRecord: 69.8ms)

After:
Completed 200 OK in 2481.0ms (Views: 2437.1ms | ActiveRecord: 5.2ms)